requirements.txt
feedparser>=6.0
requests>=2.28
beautifulsoup4>=4.12
pyyaml>=6.0
python-dotenv>=1.0
